+++ src/api/s3/s3.controller.ts
@@ -11,7 +11,7 @@ import { S3Service } from './s3.service';
 
 @Controller('s3')
 export class S3Controller {
-  constructor(private readonly fileService: S3Service) {}
+  constructor(private readonly s3Service: S3Service) {}
 
   @Post('upload')
   @UseInterceptors(FileInterceptor('file'))
@@ -19,6 +19,6 @@ export class S3Controller {
     @UploadedFile() image: Express.Multer.File,
     @Body('path') path?: string,
   ) {
-    return this.fileService.uploadFile(image, path);
+    return this.s3Service.uploadFile(image, path);
   }
 }
