+++ src/api/s3/s3.service.ts
@@ -1,56 +1,21 @@
-import {
-  S3Client,
-  PutObjectCommand,
-  DeleteObjectCommand,
-} from '@aws-sdk/client-s3';
 import { Injectable } from '@nestjs/common';
-import { ConfigService } from '@nestjs/config';
 import { v4 as uuid } from 'uuid';
+
+import { S3Repo } from '@/repository/s3.repo';
 
 @Injectable()
 export class S3Service {
-  s3Client: S3Client;
-
-  constructor(private readonly configService: ConfigService) {
-    this.s3Client = new S3Client({
-      region: configService.get('AWS_BUCKET_REGION'),
-    });
-  }
+  constructor(private readonly s3Repo: S3Repo) {}
 
   async uploadFile(file: Express.Multer.File, path = 'post/') {
     path = path.endsWith('/') ? path : path + '/';
-
-    try {
-      const bucket = this.configService.get('AWS_BUCKET_NAME');
-      const ext = file.mimetype.split('/')[1];
-      const key = path + uuid() + `.${ext}`;
-      const putCommand = new PutObjectCommand({
-        Bucket: bucket,
-        Key: key,
-        Body: file.buffer,
-      });
-      const [_, region] = await Promise.all([
-        this.s3Client.send(putCommand),
-        this.s3Client.config.region(),
-      ]);
-
-      return `https://${bucket}.s3.${region}.amazonaws.com/${key}`;
-    } catch (e) {
-      throw new Error(e);
-    }
+    const ext = file.mimetype.split('/')[1];
+    const key = path + uuid() + `.${ext}`;
+    return this.s3Repo.uploadFile(key, file.buffer);
   }
 
   async deleteFile(imagePath: string) {
     const key = imagePath.split('.com/')[1];
-    try {
-      const bucket = this.configService.get('AWS_BUCKET_NAME');
-      const deleteCommand = new DeleteObjectCommand({
-        Bucket: bucket,
-        Key: key,
-      });
-      await this.s3Client.send(deleteCommand);
-    } catch (e) {
-      throw new Error(e);
-    }
+    return this.s3Repo.deleteFile(key);
   }
 }
